Split API provider adapters
