+++ service-worker.js
@@ -91,6 +91,14 @@ const APP_SHELL = [
   '/js/api-provider-storage-runtime.js',
   '/js/api-provider-storage.js',
   '/js/api-transport.js',
+  '/js/api-openai-compatible.js',
+  '/js/api-local.js',
+  '/js/api-venice.js',
+  '/js/api-openrouter.js',
+  '/js/api-openrouter-oauth.js',
+  '/js/api-routstr.js',
+  '/js/api-ppq.js',
+  '/js/api-custom.js',
   '/js/startup-foundation.js',
   '/js/startup-maintenance-runtime.js',
   '/js/startup-profile.js',
