+++ tests/test-quality-guardrails.js
@@ -79,6 +79,14 @@ assert('CI enforces quality guardrails',
     testWorkflowSrc.includes('run: npm run quality'));
 const highValueCheckJsModules = [
   'js/api.js',
+  'js/api-custom.js',
+  'js/api-local.js',
+  'js/api-openai-compatible.js',
+  'js/api-openrouter-oauth.js',
+  'js/api-openrouter.js',
+  'js/api-ppq.js',
+  'js/api-routstr.js',
+  'js/api-venice.js',
   'js/client-list-runtime.js',
   'js/client-list.js',
   'js/dashboard-lab-widget-renderers.js',
