+++ tests/verify-modules.js
@@ -106,7 +106,19 @@
     '/js/provider-model-controls-runtime.js',
     '/js/provider-panel-renderers-runtime.js',
     '/js/api-runtime.js',
+    '/js/api.js',
+    '/js/api-models.js',
     '/js/api-provider-storage-runtime.js',
+    '/js/api-provider-storage.js',
+    '/js/api-transport.js',
+    '/js/api-openai-compatible.js',
+    '/js/api-local.js',
+    '/js/api-venice.js',
+    '/js/api-openrouter.js',
+    '/js/api-openrouter-oauth.js',
+    '/js/api-routstr.js',
+    '/js/api-ppq.js',
+    '/js/api-custom.js',
     '/js/charts-runtime.js',
     '/js/notes-runtime.js',
     '/js/pdf-import-review-runtime.js',
