+++ tsconfig.checkjs.json
@@ -7,6 +7,14 @@
   "include": [
     "types/globals.d.ts",
     "js/api.js",
+    "js/api-openai-compatible.js",
+    "js/api-local.js",
+    "js/api-venice.js",
+    "js/api-openrouter.js",
+    "js/api-openrouter-oauth.js",
+    "js/api-routstr.js",
+    "js/api-ppq.js",
+    "js/api-custom.js",
     "js/app-ai-interaction-modules.js",
     "js/app-data-io-modules.js",
     "js/app-event-listeners.js",
