+++ tsconfig.json
@@ -201,6 +201,14 @@
     "js/api-provider-storage-runtime.js",
     "js/api-provider-storage.js",
     "js/api-transport.js",
+    "js/api-openai-compatible.js",
+    "js/api-local.js",
+    "js/api-venice.js",
+    "js/api-openrouter.js",
+    "js/api-openrouter-oauth.js",
+    "js/api-routstr.js",
+    "js/api-ppq.js",
+    "js/api-custom.js",
     "js/provider-local-ai-runtime.js",
     "js/provider-local-ai-controls.js",
     "js/provider-model-controls-runtime.js",
